pyproject.toml
[project]
name = "world-forecasting"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.13"
dependencies = [
    "accelerate>=1.10.1",
    "datasets>=4.0.0",
    "docker>=7.1.0",
    "dspy>=3.0.3",
    "fasttext>=0.9.3",
    "fire>=0.7.1",
    "gradio-client>=1.12.1",
    "human-readable>=2.0.0",
    "ipython>=9.5.0",
    "more-itertools>=10.7.0",
    "openai>=1.102.0",
    "plumbum>=1.9.0",
    "pycountry>=24.6.1",
    "pyrate-limiter==2.8.0",
    "pandas>=2.2.3",
    "openpyxl>=3.1.5",
    "python-dotenv>=1.1.1",
    "requests-cache>=1.2.1",
    "requests-ratelimiter>=0.4.0",
    "sentence-transformers>=5.1.0",
    "torch>=2.8.0",
    "transformers>=4.56.0",
    "xformers>=0.0.32.post2",
    "python-slugify>=8.0.4",
    "fastapi>=0.128.0",
    "wikipedia>=1.4.0",
    "markdownify>=1.2.2",
    "sqlitedict>=2.1.0",
    "uvicorn>=0.40.0",
    "numpy>=2.3.2",
    "bm25s>=0.2.14",
    "numba>=0.63.1",
    "arrow-tokenize>=1.0.2",
    "luxical>=1.1.2",
    "leann>=0.3.6",
    "faiss-cpu>=1.13.2",
    "fastmcp>=2.14.3",
    "mxbai-rerank>=0.1.6",
    "mlflow>=3.8.1",
    "awscli>=1.43.5",
    "boto3>=1.41.5",
]

[project.scripts]
newstalk_mcp = "src.mcp_server:main"
